Week 6 - Course Project Part 2
This week contained the addition of an Account class, which will later
be made into a superclass for checking and savings accounts. For now,
each customer has 1 account with a type of either checking or savings.
It would also turn out that I was a little too eager on performing data
validation, as the instructions for this week included the addition of
try/catch blocks for that very purpose. I'm not a huge fan of this
approach to validating input, but I don't suppose my original solution
was that much cleaner.
As I learned last week, I was actually a week early in part 1 when I
used a list. I'm not sure if/when we're supposed to learn how to use a
hashmap, but I feel it would be the best option in a realistic scenario
as it:
1. Provides easy methods to ensure all entries have a unique ID
2. In an ideal environment, provides a ludicrous get speed
3. Allows the use of account numbers that aren't simply sequential IDs
Ensuring that every account has a unique ID also makes it easier for a
customer to have multiple accounts - as many real bank customers would.
Overall, a successful week.

+++ courseProject/src/courseProject/DataEntry.java
@@ -1,4 +1,4 @@
-//Caleb Barbee CIS407 Week 4: Course Project
+//Caleb Barbee CIS407 Week 6: Course Project
 package courseProject;
 
 public class DataEntry {
@@ -28,7 +28,7 @@ public class DataEntry {
 		return (min <= entry && entry <= max);
 	}
 	
-	public static boolean validateDecimalRange(float entry, float min, float max) {
+	public static boolean validateDecimalRange(double entry, double min, double max) {
 		return (min <= entry && entry <= max);
 	}
